Clean up TODOs and other leftovers from PR #175
- Issues filed for TODOs so they won't be lost.
- Test utilities moved out of mixins and into standalone objects.
