+++ src/it/scala/com/mesosphere/cosmos/IntegrationSpec.scala
@@ -7,9 +7,6 @@ import com.twitter.finagle.http.{Request, Response}
 import io.finch.test.ServiceIntegrationSuite
 import org.scalatest.fixture
 
-import com.mesosphere.cosmos.repository.PackageSourcesStorage
-import com.mesosphere.cosmos.repository.Repository
-
 abstract class IntegrationSpec
   extends fixture.FlatSpec
   with ServiceIntegrationSuite
@@ -17,7 +14,7 @@ abstract class IntegrationSpec
 
   def createService: Service[Request, Response] = {
     val marathonPackageRunner = new MarathonPackageRunner(adminRouter)
-    val sourcesStorage = PackageSourcesStorage.constUniverse(universeUri)
+    val sourcesStorage = IntegrationTests.constUniverse(universeUri)
 
     val tempDir = Files.createTempDirectory("cosmos-integration")
     tempDir.toFile.deleteOnExit
