+++ src/main/scala/com/mesosphere/cosmos/repository/PackageSourcesStorage.scala
@@ -14,25 +14,3 @@ private[cosmos] trait PackageSourcesStorage {
 
   def delete(name: Option[String], uri: Option[Uri]): Future[List[PackageRepository]]
 }
-
-private[cosmos] object PackageSourcesStorage {
-
-  private[cosmos] def constUniverse(uri: Uri): PackageSourcesStorage = new PackageSourcesStorage {
-
-    val read: Future[List[PackageRepository]] = {
-      Future.value(List(PackageRepository("Universe", uri)))
-    }
-
-    val readCache: Future[List[PackageRepository]] = {
-      Future.value(List(PackageRepository("Universe", uri)))
-    }
-
-    def add(index: Int, packageRepository: PackageRepository): Future[List[PackageRepository]] = {
-      Future.exception(new UnsupportedOperationException)
-    }
-
-    def delete(name: Option[String], uri: Option[Uri]): Future[List[PackageRepository]] = {
-      Future.exception(new UnsupportedOperationException)
-    }
-  }
-}
